add module for easier Spring boot integration
my first attempt to create a spring boot launcher and
autoconfiguration, but we do not really provide more than a runtime
(resteasy and parson)
