+++ pom.xml
@@ -58,6 +58,9 @@
     <modules>
         <module>runtime</module>
         <module>boundary</module>
+        <module>tck</module>
+        <module>spring-integration-autoconfigure</module>
+        <module>spring-integration-starter</module>
         <module>sample</module>
     </modules>
 
@@ -92,6 +95,12 @@
             </dependency>
 
             <dependency>
+                <groupId>io.rest-assured</groupId>
+                <artifactId>rest-assured</artifactId>
+                <version>5.3.0</version>
+            </dependency>
+
+            <dependency>
                 <groupId>jakarta.json</groupId>
                 <artifactId>jakarta.json-api</artifactId>
                 <version>2.1.1</version>
@@ -105,6 +114,11 @@
                 <groupId>jakarta.inject</groupId>
                 <artifactId>jakarta.inject-api</artifactId>
                 <version>2.0.1</version>
+            </dependency>
+            <dependency>
+                <groupId>jakarta.enterprise</groupId>
+                <artifactId>jakarta.enterprise.cdi-api</artifactId>
+                <version>4.0.1</version>
             </dependency>
 
             <dependency>
