--- sample/payara/src/main/java/io/github/sebastiantoepfer/json/rpc/payara/SampleAppConfig.java
@@ -1,41 +1,0 @@
-/*
- * The MIT License
- *
- * Copyright 2023 sebastian.
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- */
-package io.github.sebastiantoepfer.json.rpc.payara;
-
-import io.github.sebastiantoepfer.json.rpc.core.Notify;
-import io.github.sebastiantoepfer.json.rpc.json.rpc.adapter.JsonRpcAdapter;
-import io.github.sebastiantoepfer.json.rpc.runtime.JsonRpcRuntime;
-import jakarta.enterprise.context.RequestScoped;
-import jakarta.enterprise.inject.Produces;
-import jakarta.inject.Singleton;
-
-@Singleton
-public class SampleAppConfig {
-
-    @Produces
-    @RequestScoped
-    public JsonRpcRuntime jsonRpcRuntime() {
-        return new JsonRpcAdapter(new Notify("hello")).createJsonRpcRuntime();
-    }
-}
