+++ spring-integration-autoconfigure/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,0 +1,1 @@
+io.github.sebastiantoepfer.json.rpc.spring.integration.JsonRpcAutoConfiguration
